# much much faster uuid checks