This patch must significantly improve UUIDs check.

There are benchmarks:
```bash
$ go test -bench=.
testing: warning: no tests to run
BenchmarkIsUUIDSimple-4    	10000000	       201 ns/op	       0 B/op	       0 allocs/op
BenchmarkIsUUID3Simple-4   	10000000	       205 ns/op	       0 B/op	       0 allocs/op
BenchmarkIsUUID4Simple-4   	10000000	       206 ns/op	       0 B/op	       0 allocs/op
BenchmarkIsUUID5Simple-4   	10000000	       201 ns/op	       0 B/op	       0 allocs/op
BenchmarkIsUUIDRx-4        	 1000000	      1020 ns/op	       0 B/op	       0 allocs/op
BenchmarkIsUUIDv3Rx-4      	 1000000	      1102 ns/op	       0 B/op	       0 allocs/op
BenchmarkIsUUIDv4Rx-4      	 1000000	      1100 ns/op	       0 B/op	       0 allocs/op
BenchmarkIsUUIDv5Rx-4      	 1000000	      1042 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/bbrodriges/bench	13.280s
```